[FLINK-36861][table-planner] CREATE TABLE AS with METADATA columns doesn't work
